🐛 confirm 흐름에서 decide_routing이 짧은 확인 문구 대신 원문 업무 요청을 보게 수정
변경 이유
- Discord engineering gateway의 confirm 단계에서 사용자가 보낸 메시지(prompt_text)는 "이대로 진행" 같은 짧은 확인 문구이고, 실제 업무 내용은 conversation 레이어가 채운 outcome.intake_prompt에 들어 있다. 그런데 routing similarity가 prompt_text를 입력으로 받으면 "이대로 진행"만 가지고 open work matching을 돌려 거의 항상 매칭 실패 → CREATE로 떨어지는 회귀가 있었다. 라우팅은 canonical 업무 내용 기준이어야 한다.

주요 변경 사항
- discord/engineering_channel_router.py: route_engineering_message가 routing_prompt = intake_prompt or prompt_text를 만들어 decide_routing(prompt=routing_prompt)에 넘긴다. intake_prompt는 outcome.intake_prompt가 비어 있으면 prompt_text로 fallback되도록 이미 348행에서 정규화되어 있어 direct-confirm/single-message 흐름은 그대로 동작. legacy continuation/new-work signal 판단은 여전히 prompt_text 기반(should_continue_existing_thread/should_start_new_thread)으로 분리되어 있어 의미가 섞이지 않는다.

- tests/test_engineering_channel_router_routing.py
  - DecideRoutingWiringTests에 _isolate_cache_for_test setUp 추가 — decide_routing이 실제 SQLite cache를 읽으므로 다른 테스트가 남긴 세션과 섞이지 않게 per-test 임시 cache로 격리
  - test_route_calls_decide_routing_in_production_path를 갱신: message.content="이대로 진행", outcome.intake_prompt="onboarding step 2 정리"일 때 decide_routing.kwargs["prompt"]가 "onboarding step 2 정리"여야 하고 "이대로 진행"이면 안 된다는 assertion (회귀 가드)
  - test_route_uses_intake_prompt_for_long_task_after_confirm 신규: 사용자가 보고한 실제 시나리오 — "Obsidian + Discord + Claude를 연결해서 개발팀이 스스로 학습하는 구조를 설계해줘" 같은 긴 업무 요청 + "이대로 진행" 메시지일 때 routing이 긴 업무 요청을 본다
  - test_route_falls_back_to_message_content_when_no_intake_prompt 신규: outcome.intake_prompt=None인 direct-confirm 경로는 message.content로 정상 fallback된다는 회귀 가드
  - 이전 RouteEngineeringMessageTests에도 동일한 cache 격리 setUp 추가

- tests/test_engineering_channel_router_forum.py: HandleResearchTurnMessageTests setUp에 reset_handled_turns_for_tests() 호출 추가 — engineering_team_runtime의 process-local _HANDLED_TURNS set이 다른 테스트에서 누적된 상태로 들어오면 handle_research_turn_message가 duplicate suppression으로 no-op 반환해서 파생 회귀를 일으키던 문제를 차단

비고
- 전체 906 tests OK (4 skipped). 이전에 901이던 시점에 한 unrelated icalendar 실패는 환경에 따라 다르게 잡혔는데 이번 실행에서는 깔끔히 통과
- intake_prompt 기반 라우팅은 confirm phrase가 짧을수록 효과가 큼 — 실제 운영 패턴과 정확히 일치
